[RecordView/FEAT] 캘린더뷰와 레코드 리스트 전환 버튼 구현(#18)
